Refactor/submission (#37)
* refactor: submission

* refactor: submission & fix register react hook form

* refactor: delete console

* feat: add submission list

* feat: responsive mobile submission list

* refactor: submission list

* add: valentine competition submission

* refactor: add FormSubmission component

* refactor: FormSubmission
