+++ components/Common/form/components/InputTextarea/InputTextarea.js
@@ -1,7 +1,13 @@
 import * as React from 'react'
-import { useFormContext } from 'react-hook-form'
 
-const InputText = ({ className = '', isError = false, name = '', ...rest }) => {
+const InputText = ({
+  className = '',
+  isError = false,
+  label,
+  register = () => {},
+  required,
+  ...rest
+}) => {
   const inputBaseStyle = `${className} input-text flex items-center relative w-full px-3 py-2 rounded-lg`
   const inputBgStyle = 'bg-comic-gray-secondary'
   const inputBorderStyle = 'outline-none '
@@ -11,9 +17,13 @@ const InputText = ({ className = '', isError = false, name = '', ...rest }) => {
     isError ? 'input-text--error' : ''
   }`
 
-  const { register } = useFormContext()
-
-  return <textarea className={inputStyle} {...register(name)} {...rest} />
+  return (
+    <textarea
+      className={inputStyle}
+      {...register(label, { required })}
+      {...rest}
+    />
+  )
 }
 
 InputText.displayName = 'InputText'
