+++ hooks/useToast.js
@@ -10,7 +10,7 @@ const ToastProvider = ({ children }) => {
     if (toastConfig.type === 'error') {
       return `text-red-600 bg-red-300 border border-red-500 rounded-md`
     } else if (toastConfig.type === 'success') {
-      return `text-green-600 bg-green-300 border border-green-500 rounded-md`
+      return `text-green-600 bg-primary border border-primary rounded-md`
     } else {
       return `bg-gray-100 rounded-md`
     }
